+++ api/src/utils/__init__.py
@@ -1,0 +1,1 @@
+from .setup_logger import setup_logger
