+++ api/src/utils/setup_logger.py
@@ -1,3 +1,4 @@
+# 標準ライブラリ
 import logging
 import os
 from datetime import datetime
